+++ AdapterUnitTests/AppLovinAdapterTests/AUTAppLovinAdapterTests.m
@@ -50,13 +50,25 @@
   NSString *testSdkKey =
       @"21345678901234567890123456789012345678901234567890123456789012345678901234567890123456";
   id appLovinSdkMock = OCMClassMock([ALSdk class]);
-  OCMExpect([appLovinSdkMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
-  OCMStub(ClassMethod([appLovinSdkMock sharedWithKey:testSdkKey settings:OCMArg.any]))
-      .andReturn(appLovinSdkMock);
-  OCMStub([appLovinSdkMock initializeSdkWithCompletionHandler:OCMArg.any])
+  OCMStub(ClassMethod([appLovinSdkMock shared])).andReturn(appLovinSdkMock);
+
+  id configMock = OCMClassMock([ALSdkInitializationConfiguration class]);
+  ALSdkInitializationConfigurationBuilder *builderMock =
+      OCMClassMock([ALSdkInitializationConfigurationBuilder class]);
+  OCMExpect([builderMock setMediationProvider:ALMediationProviderAdMob]);
+  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
+  OCMStub(ClassMethod([configMock configurationWithSdkKey:testSdkKey builderBlock:OCMOCK_ANY]))
+      .andDo(^(NSInvocation *invocation) {
+        __unsafe_unretained void (^block)(ALSdkInitializationConfigurationBuilder *builder);
+        [invocation getArgument:&block atIndex:3];
+        block(builderMock);
+      })
+      .andReturn(configMock);
+
+  OCMStub([appLovinSdkMock initializeWithConfiguration:OCMOCK_ANY completionHandler:OCMOCK_ANY])
       .andDo(^(NSInvocation *invocation) {
         __unsafe_unretained void (^completionHandler)(ALSdkConfiguration *configuration);
-        [invocation getArgument:&completionHandler atIndex:2];
+        [invocation getArgument:&completionHandler atIndex:3];
         completionHandler(nil);
       });
 
@@ -64,6 +76,7 @@
   credentials.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey};
   AUTKWaitAndAssertAdapterSetUpWithCredentials([GADMediationAdapterAppLovin class], credentials);
   OCMVerifyAll(appLovinSdkMock);
+  OCMVerifyAll(builderMock);
 }
 
 - (void)testSetUpWithMultipleSdkKeys {
@@ -75,13 +88,29 @@
       @"01234567890123456789012345678901234567890123456789012345678901234567890123456789012345";
 
   id appLovinSdkMock = OCMClassMock([ALSdk class]);
-  OCMExpect([appLovinSdkMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
-  OCMStub(ClassMethod([appLovinSdkMock sharedWithKey:OCMArg.any settings:OCMArg.any]))
-      .andReturn(appLovinSdkMock);
-  OCMStub([appLovinSdkMock initializeSdkWithCompletionHandler:OCMArg.any])
+  OCMStub(ClassMethod([appLovinSdkMock shared])).andReturn(appLovinSdkMock);
+
+  id configMock = OCMClassMock([ALSdkInitializationConfiguration class]);
+  ALSdkInitializationConfigurationBuilder *builderMock =
+      OCMClassMock([ALSdkInitializationConfigurationBuilder class]);
+  OCMExpect([builderMock setMediationProvider:ALMediationProviderAdMob]);
+  OCMExpect([builderMock setPluginVersion:GADMAdapterAppLovinAdapterVersion]);
+  OCMStub(ClassMethod([configMock configurationWithSdkKey:[OCMArg checkWithBlock:^BOOL(id obj) {
+                                    return [testSdkKey1 isEqualToString:obj] ||
+                                           [testSdkKey2 isEqualToString:obj];
+                                  }]
+                                             builderBlock:OCMOCK_ANY]))
+      .andDo(^(NSInvocation *invocation) {
+        __unsafe_unretained void (^block)(ALSdkInitializationConfigurationBuilder *builder);
+        [invocation getArgument:&block atIndex:3];
+        block(builderMock);
+      })
+      .andReturn(configMock);
+
+  OCMStub([appLovinSdkMock initializeWithConfiguration:OCMOCK_ANY completionHandler:OCMOCK_ANY])
       .andDo(^(NSInvocation *invocation) {
         __unsafe_unretained void (^completionHandler)(ALSdkConfiguration *configuration);
-        [invocation getArgument:&completionHandler atIndex:2];
+        [invocation getArgument:&completionHandler atIndex:3];
         completionHandler(nil);
       });
 
@@ -92,7 +121,8 @@
   credentials2.settings = @{GADMAdapterAppLovinSDKKey : testSdkKey2};
   AUTKWaitAndAssertAdapterSetUpWithCredentials([GADMediationAdapterAppLovin class], credentials2);
 
-  OCMVerify(times(2), [appLovinSdkMock initializeSdkWithCompletionHandler:OCMArg.any]);
+  OCMVerify(times(2), [appLovinSdkMock initializeWithConfiguration:OCMOCK_ANY
+                                                 completionHandler:OCMOCK_ANY]);
 }
 
 - (void)testSetUpFailureIfUserIsTaggedAsChild {
@@ -141,6 +171,7 @@
   GADMediationAdapterAppLovin *adapter = [[GADMediationAdapterAppLovin alloc] init];
   id appLovinSdkMock = OCMClassMock([ALSdk class]);
   id adServiceMock = OCMClassMock([ALAdService class]);
+  OCMStub(ClassMethod([appLovinSdkMock shared])).andReturn(appLovinSdkMock);
   OCMStub([appLovinSdkMock adService]).andReturn(adServiceMock);
   OCMStub([adServiceMock bidToken]).andReturn(@"token");
 
@@ -159,10 +190,6 @@
 
   credentials.settings = @{@"sdkKey" : sdkKey};
 
-  OCMStub(ClassMethod([appLovinSdkMock sharedWithKey:sdkKey
-                                            settings:GADMediationAdapterAppLovin.SDKSettings]))
-      .andReturn(appLovinSdkMock);
-
   [adapter
       collectSignalsForRequestParameters:parameters
                        completionHandler:^(NSString *_Nullable signals, NSError *_Nullable error) {
@@ -178,6 +205,7 @@
   GADMediationAdapterAppLovin *adapter = [[GADMediationAdapterAppLovin alloc] init];
   id appLovinSdkMock = OCMClassMock([ALSdk class]);
   id adServiceMock = OCMClassMock([ALAdService class]);
+  OCMStub(ClassMethod([appLovinSdkMock shared])).andReturn(appLovinSdkMock);
   OCMStub([appLovinSdkMock adService]).andReturn(adServiceMock);
   OCMStub([adServiceMock bidToken]).andReturn(@"");
 
@@ -194,10 +222,6 @@
   NSString *sdkKey =
       @"12345678901234567890123456789012345678901234567890123456789012345678901234567890123456";
   credentials.settings = @{@"sdkKey" : sdkKey};
-
-  OCMStub(ClassMethod([appLovinSdkMock sharedWithKey:sdkKey
-                                            settings:GADMediationAdapterAppLovin.SDKSettings]))
-      .andReturn(appLovinSdkMock);
 
   [adapter
       collectSignalsForRequestParameters:parameters
@@ -267,61 +291,6 @@
                          XCTAssertEqual(error.code, GADMAdapterAppLovinErrorChildUser);
                          [signalsExpectation fulfill];
                        }];
-  [self waitForExpectations:@[ signalsExpectation ]];
-}
-
-- (void)testCollectSignalsNilSdkKey {
-  GADMediationAdapterAppLovin *adapter = [[GADMediationAdapterAppLovin alloc] init];
-
-  XCTestExpectation *signalsExpectation = [[XCTestExpectation alloc] init];
-
-  AUTKRTBRequestParameters *parameters = [[AUTKRTBRequestParameters alloc] init];
-  AUTKRTBMediationSignalsConfiguration *configuration =
-      [[AUTKRTBMediationSignalsConfiguration alloc] init];
-  parameters.configuration = configuration;
-  AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
-  configuration.credentials = @[ credentials ];
-  credentials.settings = @{};
-
-  [adapter
-      collectSignalsForRequestParameters:parameters
-                       completionHandler:^(NSString *_Nullable signals, NSError *_Nullable error) {
-                         XCTAssertNil(signals);
-                         XCTAssertTrue(error.description.length > 0);
-                         XCTAssertEqual(error.code,
-                                        GADMAdapterAppLovinErrorInvalidServerParameters);
-                         [signalsExpectation fulfill];
-                       }];
-
-  [self waitForExpectations:@[ signalsExpectation ]];
-}
-
-- (void)testCollectSignalsInvalidSdkKey {
-  GADMediationAdapterAppLovin *adapter = [[GADMediationAdapterAppLovin alloc] init];
-
-  XCTestExpectation *signalsExpectation = [[XCTestExpectation alloc] init];
-
-  AUTKRTBRequestParameters *parameters = [[AUTKRTBRequestParameters alloc] init];
-  AUTKRTBMediationSignalsConfiguration *configuration =
-      [[AUTKRTBMediationSignalsConfiguration alloc] init];
-  parameters.configuration = configuration;
-  AUTKMediationCredentials *credentials = [[AUTKMediationCredentials alloc] init];
-  configuration.credentials = @[ credentials ];
-
-  NSString *sdkKey = @"123";
-
-  credentials.settings = @{@"sdkKey" : sdkKey};
-
-  [adapter
-      collectSignalsForRequestParameters:parameters
-                       completionHandler:^(NSString *_Nullable signals, NSError *_Nullable error) {
-                         XCTAssertNil(signals);
-                         XCTAssertTrue(error.description.length > 0);
-                         XCTAssertEqual(error.code,
-                                        GADMAdapterAppLovinErrorInvalidServerParameters);
-                         [signalsExpectation fulfill];
-                       }];
-
   [self waitForExpectations:@[ signalsExpectation ]];
 }