+++ adapters/AppLovin/AppLovinAdapter/Bidding/GADMRTBAdapterAppLovinInterstitialRenderer.m
@@ -27,9 +27,6 @@
   /// Data used to render an interstitial ad.
   GADMediationInterstitialAdConfiguration *_adConfiguration;
 
-  /// Instance of the AppLovin SDK.
-  ALSdk *_SDK;
-
   /// AppLovin interstitial object used to load an ad.
   ALInterstitialAd *_interstitialAd;
 }
@@ -61,24 +58,16 @@
 }
 
 - (void)loadAd {
-  NSString *SDKKey = [GADMAdapterAppLovinUtils
-      retrieveSDKKeyFromCredentials:_adConfiguration.credentials.settings];
-  if (!SDKKey) {
+  if (!ALSdk.shared) {
     NSError *error = GADMAdapterAppLovinErrorWithCodeAndDescription(
-        GADMAdapterAppLovinErrorInvalidServerParameters, @"Invalid server parameters.");
-    _adLoadCompletionHandler(nil, error);
-    return;
-  }
-
-  _SDK = [GADMAdapterAppLovinUtils retrieveSDKFromSDKKey:SDKKey];
-  if (!_SDK) {
-    NSError *error = GADMAdapterAppLovinNilSDKError(SDKKey);
+        GADMAdapterAppLovinErrorAppLovinSDKNotInitialized,
+        @"Failed to retrieve ALSdk shared instance. ");
     _adLoadCompletionHandler(nil, error);
     return;
   }
 
   // Create interstitial object.
-  _interstitialAd = [[ALInterstitialAd alloc] initWithSdk:_SDK];
+  _interstitialAd = [[ALInterstitialAd alloc] initWithSdk:ALSdk.shared];
   [_interstitialAd setExtraInfoForKey:@"google_watermark" value:_adConfiguration.watermark];
 
   GADMAppLovinRTBInterstitialDelegate *delegate =
@@ -87,7 +76,7 @@
   _interstitialAd.adVideoPlaybackDelegate = delegate;
 
   // Load ad.
-  [_SDK.adService loadNextAdForAdToken:_adConfiguration.bidResponse andNotify:delegate];
+  [ALSdk.shared.adService loadNextAdForAdToken:_adConfiguration.bidResponse andNotify:delegate];
 }
 
 #pragma mark - GADMediationInterstitialAd
@@ -95,7 +84,7 @@
 - (void)presentFromViewController:(UIViewController *)viewController {
   // Update mute state
   GADMAdapterAppLovinExtras *extras = _adConfiguration.extras;
-  _SDK.settings.muted = extras.muteAudio;
+  ALSdk.shared.settings.muted = extras.muteAudio;
 
   [_interstitialAd showAd:_ad];
 }